requirements.txt
gradio==4.44.1
httpx==0.27.0
python-dotenv==1.0.1
typing-extensions==4.9.0
aiosqlite==0.19.0
tenacity==8.2.3
pydantic>=2.8.0
pydantic>=2.30